5.12. Модуляция.
