+++ js/data.js
@@ -1,3 +1,5 @@
+import {getRandomArrayElement, getRandomInteger} from './util';
+
 const ADVERTISEMENT_COUNT = 10;
 
 const LAT_LIMIT_MIN = 35.65000;
@@ -17,4 +19,33 @@ const PHOTOS = [
   'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
 ];
 
-export {ADVERTISEMENT_COUNT, LAT_LIMIT_MIN, LAT_LIMIT_MAX, LNG_LIMIT_MIN, LNG_LIMIT_MAX, FEATURES, PHOTOS, TIMES, TYPES};
+const getRandomAvatar = () => {
+  const referenceNumber = getRandomArrayElement(ADVERTISEMENT_COUNT);
+  return referenceNumber < 10 ? `img/avatars/user0${referenceNumber}.png`: `img/avatars/user${referenceNumber}.png`;
+};
+
+const getRandomLocationPoint = (min, max, decimalPlaces) => {
+  if (min >= max) {
+    throw new Error('Недопустимое значение диапазона');
+  }
+  return (Math.random() * (max - min) + min).toFixed(decimalPlaces);
+};
+
+
+const getRandomArrayElements = (items) => {
+  const maxLength = items.length;
+  const lengthOfArray = getRandomInteger(1, maxLength);
+  const elements = [];
+
+  while (elements.length < lengthOfArray) {
+    const indexOfEl = getRandomInteger(0, maxLength - 1);
+    const el = items[indexOfEl];
+
+    if (!elements.includes(el)) {
+      elements.push(el);
+    }
+  }
+  return elements;
+};
+
+export {ADVERTISEMENT_COUNT, LAT_LIMIT_MIN, LAT_LIMIT_MAX, LNG_LIMIT_MIN, LNG_LIMIT_MAX, FEATURES, PHOTOS, TIMES, TYPES, getRandomArrayElements, getRandomLocationPoint, getRandomAvatar };
