+++ js/main.js
@@ -3,4 +3,5 @@ import {ADVERTISEMENT_COUNT} from './data.js';
 
 const advertisements = new Array(ADVERTISEMENT_COUNT).fill(null).map(createAdvertisement);
 
+// eslint-disable-next-line no-unused-expressions
 advertisements;
